gptel-backend
This is all generated code that has been lightly tested. It likely has
lots of issues and needs to be able to do more. But for now I will be
using this while I develop and improve it as I find issues.

Why this instead of emacs-cody? Personally I prefer chat over
autocomplete. gptel provides fantastic hooks all over emacs for
chatting. I also see this as a useful backend to include our smart
context in the future.